Kick off
This is the initial version of the SDK.

So far, we generated the code with aas-core-codegen 0.0.14 based on
aas-core-meta 2022.6.19. The test data has been generated with
aas-core3.0rc02-testgen.

The test data involves only JSON data. The tests with XML are still
pending. The functions on the structures have not been tested (such as
``Descend``, ``DescendOnce`` or ``XorDefault``).