Operate on interfaces instead of concrete classes
We operate on *interfaces* instead of concrete classes to allow for
custom extensions and wrappers around our model classes.

Originally, we used type overloading to dispatch the visit calls. After
we decided to support custom wrappers and enhancements to our classes,
we had to switch here to interfaces instead of concrete classes.
The type overloading does not work anymore in this setting, as
descendants of *concrete* classes would be wrongly dispatched. That is
why we now dispatch explicitly, by having different visit method names.