Introduce enhancing
We want to allow downstream users to "enhance", *i.e.* wrap, our model
classes with custom properties and methods.

To that end, we generate as much boiler plate code as possible to make
their wrapping task easier.

In this patch, we also fix various omissions from the previous commit
(#86) in ``types.cs``. Namely:

* we did not generate the complete interfaces (the methods were
  missing), and
* the C# snippets were not adapted yet.

These omissions were only revealed once we started testing the
enhancing.

At this point, it becomes too cumbersome to maintain the
double-documentation on dotnet fiddle, so we only keep the examples in
our docs and do not update the fiddles anymore.